# api/docs/swagger.yaml
definitions:
  models.Animal:
    properties:
      animal_type_id:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      disease:
        type: string
      gender:
        type: string
      id:
        type: string
      last_fed_time:
        type: string
      last_watered_time:
        type: string
      name:
        type: string
      updated_at:
        type: string
      weight:
        type: integer
    type: object
  models.AnimalType:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      feeding_interval:
        type: integer
      id:
        type: string
      type:
        type: string
      updated_at:
        type: string
      watering_interval:
        type: integer
    type: object
  models.AnimalTypeCreate:
    properties:
      feeding_interval:
        type: integer
      type:
        type: string
      watering_interval:
        type: integer
    type: object
  models.AnimalsCreate:
    properties:
      animal_type_id:
        type: string
      disease:
        type: string
      gender:
        type: string
      last_fed_time:
        type: string
      last_watered_time:
        type: string
      name:
        type: string
      weight:
        type: integer
    type: object
...
  models.ListAnimalTypeRes:
    properties:
      animal_types:
        items:
          $ref: '#/definitions/models.AnimalType'
        type: array
      count:
        type: integer
    type: object
  models.ListAnimalsRes:
    properties:
      animals:
        items:
          $ref: '#/definitions/models.Animal'
        type: array
      count:
        type: integer
    type: object
...
  models.ResponseError:
    properties:
      data:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  models.StatusRes:
    properties:
      status:
        type: boolean
    type: object
  models.UpdateAnimalReq:
    properties:
      disease:
        type: string
      gender:
        type: string
      id:
        type: string
      last_fed_time:
        type: string
      last_watered_time:
        type: string
      name:
        type: string
      weight:
        type: integer
    type: object
  models.UpdateAnimalTypeReq:
    properties:
      feeding_interval:
        type: integer
      id:
        type: string
      type:
        type: string
      watering_interval:
        type: integer
    type: object
...
host: localhost:9050
info:
  contact: {}
  title: CRM Farmish
  version: "1.7"
paths:
  /v1/animal-type:
    delete:
      consumes:
      - application/json
      description: DeleteAnimalTypes - Api for Delete Animal Types
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StatusRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Delete Animal Types
      tags:
      - Animal Types
    get:
      consumes:
      - application/json
      description: ListAnimalTypes - Api for List Animal Types
      parameters:
      - in: query
        name: delete_status
        type: boolean
      - in: query
        name: limit
        type: integer
      - in: query
        name: order_by
        type: string
      - in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ListAnimalTypeRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Lis tAnimal Types
      tags:
      - Animal Types
    post:
      consumes:
      - application/json
      description: CreateAnimalTypes - Api for crete Animal Types
      parameters:
      - description: AnimalTypeCreate
        in: body
        name: AnimalTypeCreate
        required: true
        schema:
          $ref: '#/definitions/models.AnimalTypeCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AnimalType'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Create Animal Types
      tags:
      - Animal Types
    put:
      consumes:
      - application/json
      description: UpdateAnimalTypes - Api for Update Animal Types
      parameters:
      - description: UpdateAnimalTypeReq
        in: body
        name: UpdateAnimalTypeReq
        required: true
        schema:
          $ref: '#/definitions/models.UpdateAnimalTypeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AnimalType'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Update Animal Types
      tags:
      - Animal Types
  /v1/animal-type/get:
    get:
      consumes:
      - application/json
      description: GetAnimalTypes - Api for Get Animal Types
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AnimalType'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Get Animal Types
      tags:
      - Animal Types
  /v1/animals:
    delete:
      consumes:
      - application/json
      description: DeleteAnimals - Api for Delete Animals
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StatusRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Delete Animals
      tags:
      - Animals
    get:
      consumes:
      - application/json
      description: ListAnimals - Api for List Animals
      parameters:
      - in: query
        name: delete_status
        type: boolean
      - in: query
        name: limit
        type: integer
      - in: query
        name: order_by
        type: string
      - in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ListAnimalsRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Lis tAnimals
      tags:
      - Animals
    post:
      consumes:
      - application/json
      description: CreateAnimals - Api for crete Animals
      parameters:
      - description: AnimalTypeCreate
        in: body
        name: AnimalTypeCreate
        required: true
        schema:
          $ref: '#/definitions/models.AnimalsCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Animal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Create Animal s
      tags:
      - Animals
    put:
      consumes:
      - application/json
      description: UpdateAnimals - Api for Update Animals
      parameters:
      - description: UpdateAnimalTypeReq
        in: body
        name: UpdateAnimalTypeReq
        required: true
        schema:
          $ref: '#/definitions/models.UpdateAnimalReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Animal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Update Animals
      tags:
      - Animals
  /v1/animals/get:
    get:
      consumes:
      - application/json
      description: GetAnimals - Api for Get Animals
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Animal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Get Animal s
      tags:
      - Animals
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"